Lego sets from 1970 through 2022.

This dataset was built using the getSets function. For working with the LEGO sets data frame this pre-built data is preferred as it will minimize the API requests. Note that the only disadvantage is that the rating and reviewCount may be out-of-date and inaccurate. The remaining variables are relatively static.

Format

A data.frame.

setID

integer; 18455 unique values

name

character; 15372 unique values

year

integer; 53 unique values

theme

character; 154 unique values

themeGroup

character; 16 unique values

subtheme

character; 901 unique values

category

character; 7 unique values

released

logical; 2 unique values

pieces

integer; 1403 unique values

minifigs

integer; 34 unique values

bricksetURL

character; 18455 unique values

rating

numeric; 31 unique values

reviewCount

integer; 64 unique values

packagingType

character; 18 unique values

availability

character; 10 unique values

agerange_min

integer; 16 unique values

US_retailPrice

numeric; 142 unique values

US_dateFirstAvailable

Date; 881 unique values

US_dateLastAvailable

Date; 2125 unique values

UK_retailPrice

numeric; 210 unique values

UK_dateFirstAvailable

Date; 825 unique values

UK_dateLastAvailable

Date; 1984 unique values

CA_retailPrice

numeric; 162 unique values

CA_dateFirstAvailable

Date; 641 unique values

CA_dateLastAvailable

Date; 1808 unique values

DE_retailPrice

numeric; 155 unique values

DE_dateFirstAvailable

Date; 428 unique values

DE_dateLastAvailable

Date; 1176 unique values

height

numeric; 241 unique values

width

numeric; 286 unique values

depth

numeric; 279 unique values

weight

numeric; 1046 unique values

thumbnailURL

character; 17420 unique values

imageURL

character; 17420 unique values

Source

https://brickset.com